requirements.txt
SQLAlchemy>=1.4.0,<2.0.0
PyMySQL>=1.0.0,<2.0.0
python-dotenv>=0.19.0,<1.0.0
